pyproject.toml
[tool.poetry]
name = "vehicle-project"
version = "0.1.0"
description = ""
authors = ["Uliana Semyonova <i@ulianasemyonova.ru>"]
readme = "README.md"
packages = [{include = "vehicle_project"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
pillow = "^10.0.1"
psycopg2 = "^2.9.9"
django-filter = "^23.3"
djangorestframework-simplejwt = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
